moved trades, make agressive fix
