+++ src/main/java/de/teamlapen/vampirism/command/test/MakeVillagerAgressiveCommand.java
@@ -3,6 +3,9 @@ package de.teamlapen.vampirism.command.test;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 
 import de.teamlapen.lib.lib.util.BasicCommand;
+import de.teamlapen.vampirism.api.entity.factions.IFactionEntity;
+import de.teamlapen.vampirism.api.entity.hunter.IHunter;
+import de.teamlapen.vampirism.api.entity.vampire.IVampire;
 import de.teamlapen.vampirism.tileentity.TotemTileEntity;
 import net.minecraft.command.CommandSource;
 import net.minecraft.command.Commands;
@@ -27,6 +30,7 @@ public class MakeVillagerAgressiveCommand extends BasicCommand {
     private static int makeVillagerAgressive(ServerPlayerEntity asPlayer) {
         List<VillagerEntity> l = asPlayer.getEntityWorld().getEntitiesWithinAABB(VillagerEntity.class, asPlayer.getBoundingBox().grow(3, 2, 3));
         for (VillagerEntity v : l) {
+            if(v instanceof IFactionEntity)continue;
             TotemTileEntity.makeAgressive(v);
         }
         return 0;
